Update installation instructions and fix options handling
The README should now have enough info for people to know how to use
this.  It's by no means complete, but good enough for a side project
that will get merged elsewhere when it's better.

When I converted to passing the args as a splat, the face/action code
tried to be helpful and added an options hash at the end, which broke my
completion.  I'm going to just manually remove that for now.  Maybe I'll
switch back to using an array later.